handle protected mmcblk_boot_ devices
Some block devices support physical write protection. The kernel
provides a standard interface to enable or disable protection in
/sys/class/block/*/force_ro.

This patch adds functionality to automatically detect these memory
types. If read-only mode is enabled on the partition on which the
uboot environment must be written, libubootenv temporarily switches
to read/write mode.

The implementation should be generic enough to add support for
other device types with differently implemented write protection
later. This is the motivation for object-oriented design.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>